Disable _FORTIFY_SOURCE auto-insertion for musl targets
This commit adjusts the build script for the `onig_sys` crate to disable
the automatic insertion of the _FORTIFY_SOURCE macro when compiling for
musl targets. By unsetting _FORTIFY_SOURCE before setting it explicitly
to 0, we avoid potential compilation issues and warnings related to
_FORTIFY_SOURCE being redefined. This change is specifically aimed at
improving compatibility and build process for projects targeting musl libc,
ensuring a smoother compilation experience without the need to manually
adjust the environment or compiler flags outside of the build script.